// src/types/enums.ts
export enum LabTestStatus {
  PENDING = 'PENDING',
  SAMPLE_COLLECTED = 'SAMPLE_COLLECTED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  REVIEWED = 'REVIEWED',
  CANCELLED = 'CANCELLED',
}

export enum LabSampleType {
  BLOOD = 'BLOOD',
  URINE = 'URINE',
  STOOL = 'STOOL',
  SPUTUM = 'SPUTUM',
  SALIVA = 'SALIVA',
  BIOPSY = 'BIOPSY',
  CSF = 'CSF',
  OTHER = 'OTHER',
}

export enum LabTestPriority {
  ROUTINE = 'ROUTINE',
  URGENT = 'URGENT',
  STAT = 'STAT',
}